==> Scripts/3-HDT_compile-databases.Rmd <==
---
title: "Compile Databases"
author: "Bari Fuchs" 
output: word_document
editor_options: 
  chunk_output_type: console
---

This script was written by Bari Fuchs for the HDT and intake study. The purpose of the script is to compile participant_data, behavioral_metrics, and VPP_parameter_means datasets into 1 dataset to be used for HDT paper analyses

```{r Import datasets, echo = TRUE}
library(plyr)

# set basedir to navigate to HDT directory (modify if on different machine)
basedir <- ("~/OneDrive - The Pennsylvania State University/b-childfoodlab_Shared/Inactive_Studies/DMK_Study/AB_reprocess/HDT/HDT-intake-git")

# set directory for generated datasets
setwd(file.path(basedir,"Data/GeneratedDatabases/"))

# Non-DM data (e.g. demographics, intake data, liking ratings)
HDT_pardat <-  read.csv("HDT_participant_data.csv", stringsAsFactors = FALSE)

# Behavioral metrics (generated by 2b-HDT_compute-behavioralmetrics.Rmd)
HDT_behavioralmetrics <-  read.csv("HDT_behavioral_metrics.csv", stringsAsFactors = FALSE)

# VPP parameter values (generated by VPP scripts)
HDT_VPP_means <-  read.csv("HDT_VPP_means.csv", stringsAsFactors = FALSE)


```

```{r Merge databases, echo = TRUE}
# Merge datases
HDT_temp <- merge(HDT_pardat, HDT_behavioralmetrics, by="subjID")
HDT_plus <- merge(HDT_temp, HDT_VPP_means, by="subjID")

```


# Export files for analyses
```{r Export, echo = TRUE}

setwd(file.path(basedir,"Data/GeneratedDatabases/"))
write.csv(HDT_plus,'HDT_plus.csv')

```
